:recycle: 주석 한글화 및 리팩토링

+++ src/main/java/backendnovice/projectbookshop/board/article/dto/ArticleDTO.java
@@ -1,8 +1,8 @@
 /**
  * @author   : backendnovice@gmail.com
  * @created  : 2023-07-25
- * @modified : 2023-09-04
- * @desc     : An article-related data transfer object class. that used to communicate between layers.
+ * @modified : 2023-09-18
+ * @desc     : 레이어 간의 통신을 담당하는 게시글 데이터 전송 객체.
  */
 
 package backendnovice.projectbookshop.board.article.dto;
@@ -33,7 +33,7 @@ public class ArticleDTO {
 
     private Page<CommentDTO> commentPages;
 
-    private int commentCount;
+    private long commentCount;
 
     private LocalDateTime date;
 }
